requirements.txt
annotated-types==0.7.0
array-api-compat==1.13.0
Box2D==2.3.10
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Farama-Notifications==0.0.4
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
gymnasium==1.2.3
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
seaborn==0.13.2
six==1.17.0
swig==4.4.1
sympy==1.14.0
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
